// src/common/enums/HttpStatusCode.ts
export enum EnumHttpStatusCode {
  OK_200 = 200,
  CREATED_201 = 201,
  DELETED_204 = 204,
  BAD_REQUEST_400 = 400,
  UNAUTHENTICATE_401 = 401,
  FORBIDDEN_403 = 403,
  NOT_FOUND_404 = 404,
  VALIDATION_ERROR_422 = 422,
  DUPLICATE_DATA_409 = 409,
  RATE_LIMITER_429 = 429,
  INTERNAL_SERVER_500 = 500,
  TIMEOUT_ERROR_504 = 504,
}